requirements.txt
pymysql
pymongo
prettytable
python-dotenv
mysql-connector-python
colorama
tabulate